fix: close files properly when BuildDockerfiles exits
Previous code called defer with argument "dockerfiles..." which is
always empty at time of evaluation.

Even when dockerfiles is appended to later in this method, it has no
entries at time of evaluation as an argument to the defer function, thus
no files were ever closed as a result.

Minor refactor so that all referenced files and HTTP requests are now
closed at end of function.

Signed-off-by: Adam Eijdenberg <adam@continusec.com>